Fix filetransfer image not known bug
This commit fixes a bug where if the clean method and
filetransfer method are being run at the same time,
the filetransfer method will sometimes fail with
an "image not known" error message.

This is due to the clean method removing all unused
images if the all option is set. This commit ensures
that the image that the filetransfer method makes
use of is available when the filetransfer method
wants to create the container it uses to deposit
the file on the host system.

Signed-off-by: Joseph Sawaya <jsawaya@redhat.com>